requirements.txt
python-dotenv~=1.0.1
discord.py~=2.5.2
rich~=13.9.4